try int floor for unhash

+++ cogs/meta.py
@@ -455,7 +455,7 @@ class Meta:
 
     def unhash_ID(self: int):
         # (id-1 / 2), swap(first 2 numbers, last 2 numbers), + E
-        user_id = (self - 1)/2
+        user_id = int((self - 1)/2)
         user_id = str(user_id)[-2:] + str(user_id)[2:-2] + str(user_id)[:2]
         user_id = int(user_id) + ord('E')
         return user_id
